requirements.txt
botcity-maestro-sdk>=0.3.3,<1.0